test: Better tests for random row validation on large decimals
