+++ tests/unit/ibis_addon/test_operations.py
@@ -182,3 +182,99 @@ def test_execute_epoch_seconds_new(
 ):
     result = module_under_test.execute_epoch_seconds_new(None, test_input)
     assert list(result) == list(expected)
+
+
+def test_bigquery_decimal_literal_compiles_without_scientific_notation():
+    import decimal
+    import ibis.expr.datatypes as dt
+    from ibis.backends.bigquery.compiler import BigQueryCompiler
+
+    schema = ibis.schema(
+        {
+            "id": dt.Decimal(76, 38),
+            "small_dec": dt.Decimal(38, 9),
+        }
+    )
+    t = ibis.table(schema, name="dvt_large_decimals")
+    large_vals = [
+        decimal.Decimal("323456789012345678901234567890"),
+        decimal.Decimal("223456789012345678901234567890"),
+    ]
+    small_vals = [
+        decimal.Decimal("123.456"),
+        decimal.Decimal("789.012"),
+    ]
+
+    sql_large = str(BigQueryCompiler().to_sql(t.filter(t.id.isin(large_vals))))
+    sql_small = str(BigQueryCompiler().to_sql(t.filter(t.small_dec.isin(small_vals))))
+
+    assert "323456789012345678901234567890" in sql_large
+    assert "BIGNUMERIC" in sql_large
+    assert "E+" not in sql_large
+
+    assert "123.456" in sql_small
+    assert "NUMERIC" in sql_small
+
+
+def test_db2_decimal_literal_compiles_without_scientific_notation():
+    import decimal
+    import ibis.expr.datatypes as dt
+    import sqlalchemy as sa
+    from third_party.ibis.ibis_db2.compiler import Db2Compiler
+
+    schema = ibis.schema({"id": dt.Decimal(31, 0)})
+    t = ibis.table(schema, name="dvt_large_decimals")
+    vals = [decimal.Decimal("223456789012345678901234567890")]
+    expr = t.filter(t.id.isin(vals))
+
+    ast = Db2Compiler.to_ast(expr.op())
+    query = ast.queries[0]
+    metadata = sa.MetaData()
+    sa_t = sa.Table("dvt_large_decimals", metadata, sa.Column("id", sa.DECIMAL(31, 0)))
+    query.context.set_ref(t.op(), sa_t)
+
+    w = query._translate(query.where[0])
+    compiled = str(w.compile(compile_kwargs={"literal_binds": True}))
+    assert "223456789012345678901234567890" in compiled
+    assert "E+" not in compiled
+
+
+def test_mssql_decimal_literal_compiles_without_scientific_notation():
+    import decimal
+    import ibis.expr.datatypes as dt
+    import sqlalchemy as sa
+    from ibis.backends.mssql.compiler import MsSqlCompiler
+
+    schema = ibis.schema({"id": dt.Decimal(38, 0)})
+    t = ibis.table(schema, name="dvt_large_decimals")
+    vals = [decimal.Decimal("223456789012345678901234567890")]
+    expr = t.filter(t.id.isin(vals))
+
+    ast = MsSqlCompiler.to_ast(expr.op())
+    query = ast.queries[0]
+    metadata = sa.MetaData()
+    sa_t = sa.Table("dvt_large_decimals", metadata, sa.Column("id", sa.DECIMAL(38, 0)))
+    query.context.set_ref(t.op(), sa_t)
+
+    w = query._translate(query.where[0])
+    compiled = str(
+        w.compile(
+            dialect=sa.dialects.mssql.dialect(), compile_kwargs={"literal_binds": True}
+        )
+    )
+    assert "223456789012345678901234567890" in compiled
+    assert "E+" not in compiled
+
+
+def test_null_decimal_literal():
+    import ibis.expr.datatypes as dt
+    from ibis.backends.mssql.compiler import MsSqlCompiler
+    from third_party.ibis.ibis_db2.compiler import Db2Compiler
+
+    expr = ibis.literal(None, type=dt.Decimal(10, 2))
+    # Test that compiling NULL decimal literal does not raise TypeError
+    mssql_res = MsSqlCompiler.to_ast(expr.op()).queries[0]._translate(expr.op())
+    db2_res = Db2Compiler.to_ast(expr.op()).queries[0]._translate(expr.op())
+
+    assert str(mssql_res) == "NULL"
+    assert str(db2_res) == "NULL"
